pkp/pkp-lib#10449 code cleanup
